requirements.txt
pandas>=2.0
numpy>=1.25
scipy>=1.11
tushare>=1.2.89
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
plotly>=5.15.0
